+++ packages/dashboard/src/features/storage/pages/BucketsPage.tsx
@@ -3,7 +3,8 @@ import { useQueryClient } from '@tanstack/react-query';
 import { Upload } from 'lucide-react';
 import PencilIcon from '../../../assets/icons/pencil.svg?react';
 import RefreshIcon from '../../../assets/icons/refresh.svg?react';
-import { useStorage } from '../hooks/useStorage';
+import { useStorageBuckets } from '../hooks/useStorageBuckets';
+import { useStorageObjects } from '../hooks/useStorageObjects';
 import { StorageSidebar } from '../components/StorageSidebar';
 import { StorageDataGrid } from '../components/StorageDataGrid';
 import { BucketFormDialog } from '../components/BucketFormDialog';
@@ -64,10 +65,9 @@ export default function BucketsPage() {
     bucketsError,
     refetchBuckets,
     useBucketStats,
-    uploadObject,
-    deleteObjects,
     deleteBucket,
-  } = useStorage();
+  } = useStorageBuckets();
+  const { uploadObject, deleteObjects } = useStorageObjects();
   const selectedBucket = useMemo(() => {
     if (isLoading || !buckets.length) {
       return null;
